double -> int 변경

+++ src/operations/Division.java
@@ -1,17 +1,17 @@
 package operations;
 
 public class Division {
-    private double firstNum;
-    private double secondNum;
+    private int firstNum;
+    private int secondNum;
 
     //세터
-    public void setDivision(double firstNum, double secondNum) {
+    public void setDivision(int firstNum, int secondNum) {
         this.firstNum = firstNum;
         this.secondNum = secondNum;
     }
 
     //게터
-    public double getDivision() {
+    public int getDivision() {
         return firstNum - secondNum;
     }
 }
